+++ src/number/between.js
@@ -6,7 +6,7 @@ import _curry3 from '../_internals/_curry3'
  * @function
  * @category Number
  * @sig Number -> Number -> Number -> Boolean
- * @description Checks to see if a number is between two other provided numbers
+ * @description Checks to see if a number is between two other provided numbers (inclusive)
  * @param {Number} min The number our value should be greater than or equal too
  * @param {Number} max The number our value should be less than or equal too
  * @param {Number} n The value to compare with